Learner account CRUD controller and service

+++ Spring/ITraining_API/src/main/java/com/example/itraining_api/repository/LearnerAccountRepository.java
@@ -6,6 +6,5 @@ import com.example.itraining_api.entity.LearnerAccount;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LearnerAccountRepository extends CrudRepository<LearnerAccount, Long> {
-
+public interface LearnerAccountRepository extends CrudRepository<LearnerAccount, Integer> {
 }
